+++ app/build.gradle
@@ -47,7 +47,7 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation project(':utilcode')
+//    implementation project(':utilcode')
     implementation project(':subutil')
 
     implementation "com.android.support:appcompat-v7:$support_version"
@@ -56,7 +56,7 @@ dependencies {
     // LeakCanary
     debugImplementation "com.squareup.leakcanary:leakcanary-android:$leakcanary_version"
     releaseImplementation "com.squareup.leakcanary:leakcanary-android-no-op:$leakcanary_version"
-//    implementation 'com.blankj:utilcode:1.16.0'
+    implementation 'com.blankj:utilcode:1.16.0'
 }
