+++ app/src/main/java/com/blankj/androidutilcode/feature/core/app/OnReleasedListener.java
@@ -1,0 +1,13 @@
+package com.blankj.androidutilcode.feature.core.app;
+
+/**
+ * <pre>
+ *     author: Blankj
+ *     blog  : http://blankj.com
+ *     time  : 2018/05/23
+ *     desc  :
+ * </pre>
+ */
+public interface OnReleasedListener {
+    void onReleased();
+}
